factory_service/pyproject.toml
[tool.poetry]
name = "factory_service"
version = "0.1.0"
description = ""
authors = ["Zaikin Valentin <zaikinvala59@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.7.0"
webargs = "^7.0.1"
flask-marshmallow = "^0.14.0"
celery = "^5.0.5"
psycopg2-binary = "^2.8.6"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
